refactor(fsmv2): delete unreachable TryingToAuthenticateState and cascading dead code (ENG-4265)
TryingToAuthenticateState explicitly documented itself as "unreachable, kept
for ENG-4265 cleanup." Zero incoming transitions from any state. Delete it,
the AuthenticateAction it dispatched, and the entire cascade of dead setters,
getters, and snapshot fields that only this code path used.

The communicator no longer manages auth — TransportWorker's StartingState
handles authentication via its own AuthenticateAction and dependencies.
SyncingState/RecoveringState operate exclusively on child health counts.

Moves AuthenticatedUUID polling in main.go from communicator to transport
observed state, since TransportWorker now owns authentication.
