+++ umh-core/pkg/fsmv2/workers/communicator/snapshot/snapshot_test.go
@@ -16,7 +16,6 @@ package snapshot_test
 
 import (
 	"testing"
-	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -35,59 +34,6 @@ var _ = Describe("CommunicatorObservedState", func() {
 
 	BeforeEach(func() {
 		observed = &snapshot.CommunicatorObservedState{}
-	})
-
-	Describe("IsTokenExpired", func() {
-		Context("when token expires in 1 hour", func() {
-			BeforeEach(func() {
-				observed.JWTExpiry = time.Now().Add(1 * time.Hour)
-			})
-
-			It("should return false (not expired)", func() {
-				expired := observed.IsTokenExpired()
-				Expect(expired).To(BeFalse(), "Fresh token should not be expired")
-			})
-		})
-
-		Context("when token expired 1 hour ago", func() {
-			BeforeEach(func() {
-				observed.JWTExpiry = time.Now().Add(-1 * time.Hour)
-			})
-
-			It("should return true (expired)", func() {
-				expired := observed.IsTokenExpired()
-				Expect(expired).To(BeTrue(), "Past expiration should be expired")
-			})
-		})
-
-		Context("when token expires in 5 minutes (within 10-minute buffer)", func() {
-			BeforeEach(func() {
-				observed.JWTExpiry = time.Now().Add(5 * time.Minute)
-			})
-
-			It("should return true (considered expired for proactive refresh)", func() {
-				expired := observed.IsTokenExpired()
-				Expect(expired).To(BeTrue(), "Token expiring soon should be considered expired")
-			})
-		})
-
-		Context("when token expires in 15 minutes (outside 10-minute buffer)", func() {
-			BeforeEach(func() {
-				observed.JWTExpiry = time.Now().Add(15 * time.Minute)
-			})
-
-			It("should return false (not yet in refresh window)", func() {
-				expired := observed.IsTokenExpired()
-				Expect(expired).To(BeFalse(), "Token with >10 minutes remaining should not be expired")
-			})
-		})
-
-		Context("when no expiration is set", func() {
-			It("should return false (zero time means no expiration tracking)", func() {
-				expired := observed.IsTokenExpired()
-				Expect(expired).To(BeFalse(), "Zero expiration time should not be considered expired")
-			})
-		})
 	})
 
 	Describe("GetConsecutiveErrors", func() {
